Tweak synchronizer

+++ mempool/src/synchronizer.rs
@@ -102,7 +102,7 @@ impl Synchronizer {
         let (message, nodes) = match coin < self.sync_bias {
             true => (
                 MempoolMessage::ShardRequest(missing, self.name),
-                self.committee.quorum_threshold(),
+                self.committee.quorum_threshold() as usize,
             ),
             false => (
                 MempoolMessage::BatchRequest(missing, self.name),
